Implement Location permissions
Updated the functionality of the CustomFAB composable in the SetupScreen.
The FAB runs a utility function in Utils that checks if the user has granted
permissions previously. This updates the permissionRequestResult in the
CustomFAB composable. At the same time, inside the CustomFAB there is a
launchEffect that observes the permissionRequestResult value so if it
changes and is true, then we navigateto the RunScreen